Move utility function to separate file
